[FEAT] AttendanceService - 입실 체크 - 퇴실 체크 - 반경 100M 이내인지 확인

+++ src/main/java/com/bitbox/user/dto/CurrentLocationDto.java
@@ -9,7 +9,7 @@ import javax.validation.constraints.NotEmpty;
 @Getter
 public class CurrentLocationDto {
     @NotEmpty(message = "필수 입력값입니다")
-    private long lat;
+    private double lat;
     @NotEmpty(message = "필수 입력값입니다")
-    private long lng;
+    private double lng;
 }
